sandbox: clean up handling of XDG_RUNTIME_DIR
Move the logic for deciding to create or bind-mount XDG_RUNTIME_DIR into
a separate function.  Instead of creating a subdirectory, do it as a new
tmpfs, just like a real XDG_RUNTIME_DIR.

Set up the XDG_RUNTIME_DIR envvar correctly: we always name it according
to the name in the sandbox (which could theoretically differ from the
host).