sandbox: add explicit Wayland support
Add a dependency on wayland-rs and use it to attempt to create a
sandboxed Wayland listener via the wp_security_context_v1_name
extension.

This extension is in "staging" and isn't particularly widely supported:
neither weston nor gnome-shell have support, but I was able to test with
sway.  In order to use the extension we need to hold on to an fd to
control the lifetime of the socket: add support to the Sandbox struct
for that.

If the extension isn't present then we fall back to simply bind mounting
the host socket into the sandbox, matching Flatpak behaviour.

Change the default sandbox permissions to include wayland (securely, if
possible) and exclude the XDG_RUNTIME_DIR.