sandbox: step up home directory handling
In case we're not sharing the home directory, always use
/home/{username} instead of the value from the host.  This requires a
bit of obnoxious work, since the value of the home directory is used in
several places and now depends on the sandbox configuration, but we
don't want to recalculate it each time.

Use a tmpfs for the case where we're not binding from the host.

Disable the home directory permission by default.

At this point, the sandbox code is still a complete work in progress,
unaudited, and not to be trusted, but on the surface, it's semi-secure:
it can run simple apps (like the GNOME games) though a secured Wayland
socket without access to the home directory.  It has no D-Bus and
forgets everything each time, though.