sandbox: add initial xdg-dbus-proxy support
We can now setup xdg-dbus-proxy without filtering.

Add an ArgsFd helper to support the `--args=fd` flag supported by bwrap
and xdg-dbus-proxy.

Also add explicit support for fd handling in the child: we now attempt
to catch any leaked CLOEXEC fds and also support explicitly setting a
list of fds non-CLOEXEC.  This sort of defines a new approach: *all* fds
in the parent should be CLOEXEC and patched after the fork.  This avoids
theoretical races if two threads wanted to spawn processes at the same
time (which isn't something that can happen yet, of course).

The CLOEXEC checking caught a bug in composefs-rs, so depend on the
fixed version from git.